Refactor: More refactoring for less renaming
