+++ pyavd_utils/_bindings/__init__.pyi
@@ -2,6 +2,6 @@
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
 
-from . import passwords, schema_store, validation
+from . import _passwords, _schema_store, _validation
 
-__all__ = ["passwords", "schema_store", "validation"]
+__all__ = ["_passwords", "_schema_store", "_validation"]
