+++ pyavd_utils/_bindings/_schema_store.pyi
