+++ pyavd_utils/_bindings/_validation.pyi
