+++ rust/python-bindings/src/schema_store/mod.rs
@@ -24,33 +24,32 @@ pub(crate) fn get_store() -> PyResult<&'static Store> {
     })
 }
 
-#[pyfunction]
-pub(crate) fn init_store_from_file(file: PathBuf) -> PyResult<()> {
-    info!("Initialize the schema store from file.");
-
-    let store = {
-        let store = Store::from_file(Some(&file)).map_err(|err| {
-            PyRuntimeError::new_err(format!(
-                "Error while loading the Schema Store from file: {err}",
-            ))
-        })?;
-        store.as_resolved().map_err(|err| {
-            PyRuntimeError::new_err(format!("Error while resolving the Schema Store: {err}",))
-        })
-    }?;
-
-    STORE.set(store).map_err(|_store| {
-        PyRuntimeError::new_err(
-            "Unable to initialize the schema store. \
-                 Initialization can only happen once, and must be done before running any validations."
-                .to_owned(),
-        )
-    }).inspect(|()| info!("Initialized the schema store from file."))
-}
-
 /// Shared schema store helpers.
-#[pyo3::pymodule(name = "schema_store")]
-pub(crate) mod schema_store_mod {
-    #[pymodule_export]
-    use super::init_store_from_file;
+#[pyo3::pymodule]
+pub(crate) mod _schema_store {
+    use super::*;
+
+    #[pyfunction]
+    pub(crate) fn init_store_from_file(file: PathBuf) -> PyResult<()> {
+        info!("Initialize the schema store from file.");
+
+        let store = {
+            let store = Store::from_file(Some(&file)).map_err(|err| {
+                PyRuntimeError::new_err(format!(
+                    "Error while loading the Schema Store from file: {err}",
+                ))
+            })?;
+            store.as_resolved().map_err(|err| {
+                PyRuntimeError::new_err(format!("Error while resolving the Schema Store: {err}",))
+            })
+        }?;
+
+        STORE.set(store).map_err(|_store| {
+            PyRuntimeError::new_err(
+                "Unable to initialize the schema store. \
+                     Initialization can only happen once, and must be done before running any validations."
+                    .to_owned(),
+            )
+        }).inspect(|()| info!("Initialized the schema store from file."))
+    }
 }
